gui/requirements.txt
pyserial>=3.5
customtkinter>=5.2.0
pillow>=10.0.0  # Required for customtkinter
